Models/inference_yolo.py
"""
YOLOv5 inference module for pneumonia detection
"""
import numpy as np
from PIL import Image
import cv2

def load_yolo_model(repo_path, weights_path):
    """Load YOLOv5 model - placeholder implementation"""
    print(f"Loading YOLOv5 model from {weights_path}")
    # Return a mock model object
    class MockYOLOModel:
        def __call__(self, image):
            # Mock detection results
            class MockResults:
                def __init__(self):
                    self.xyxy = [np.array([[100, 100, 200, 200, 0.8, 0]])]  # Mock bounding box
            return MockResults()
    
    return MockYOLOModel()

def predict_yolo_image(model, image, threshold=0.5):
    """Predict using YOLOv5 model - placeholder implementation"""
    # Convert PIL to numpy if needed
    if isinstance(image, Image.Image):
        image_np = np.array(image)
    else:
        image_np = image
    
    # Mock prediction
    prediction = "Normal"  # Default prediction
    probability = 0.3  # Mock probability
    
    # If probability > threshold, predict pneumonia
    if probability > threshold:
        prediction = "Pneumonia"
        # Draw mock bounding box
        image_np = cv2.rectangle(image_np, (100, 100), (200, 200), (255, 0, 0), 2)
        image_np = cv2.putText(image_np, f"Pneumonia {probability:.2f}", 
                              (100, 90), cv2.FONT_HERSHEY_SIMPLEX, 0.6, (255, 0, 0), 2)
    
    # Convert back to PIL
    annotated_image = Image.fromarray(image_np)
    
    return annotated_image, prediction, probability